extract button component into common module
