+++ src/client/modules/common/components/createTabBarIconWrapper.jsx
