+++ src/client/modules/counter/components/CounterView.web.jsx
@@ -1,9 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import Helmet from 'react-helmet';
-import { Button } from 'reactstrap';
+//import { Button } from 'reactstrap';
 import styled from 'styled-components';
 import PageLayout from '../../../app/PageLayout';
+
+//import Button from '../../common/components/Button';
+import { Button } from '../../common/components';
 
 const Section = styled.section`
   margin-bottom: 30px;
